Always send task stack change updates if PiP is enabled
PiP stack repositioning is triggered when task stack change
is detected. Usually these updates are sent when there was
a visibility change since last pause. In case when we're
navigating from setting back to launcher on TV there is no
visibility change. This happened to work in most cases
because PipOverlayActivity was shown or hidden somewhere
between user actions.

This CL always notifies about task stack changes if pinned
stack is present on activity pause.

Bug: 29344059
Change-Id: I5ab3f2b8a97983fcb79d58626b6494f8ec31dbf3

Former-commit-id: 8290f8f66a75e89ac3ef4d399798486e7d8682a8