Update DocumentsUI to use theme colors.
Properly assigns the app's colors for primary, primaryDark, accent, etc
to the colors used by the framework.

Moves a couple colors that are alternate shades of the accent color
into the framework for ease of overlaying.

Deletes a few unused colors as well.

Finally, fixes a bug where colorControlActivated was not being
properly set due to the theme depending upon AppCompat. To
workaround this undefined behavior, we explicitly set
colorControlActivated to colorAccent which is what the framework
sets it to.

Change-Id: Ie0cba19b1b83a21f65692102df91bf4daf27dc44
Fixes: 29581883

Former-commit-id: 0e59a22fccbfe277daa4e3b3f35b6c5e15668ba6