CEC: Change the logical address allocation policy
According to the CEC spec (Section 7), a message transmitted and
acknowledged should be assumed correctly received. Therefore, If a
polling message to certain logical address(LA) is acknowledged during
the LA allocation phase, it can be assumed that there exists a CEC
device on the LA. This CL will see whether at least 1 polling message is
acknowledged or not in order to check the CEC device existence, instead
of using the previous voting policy.

Bug: 27337386
Change-Id: I39df78f6db29a1b47cb962f8fe5e0382c16028de

Former-commit-id: dbf9f9cc5e363f26f8dfc4ceef6c6abb3881a190