Camera2: Add @hide variant for openCamera for UID forwarding
Trusted services may open a camera device on behalf of some client
of theirs; such services need to forward the UID of their client to
the camera service for validation of permissions, etc.

Add a variant of openCamera that makes this simple, only accessible
to unbundled services for now.  Only services explicitly trusted
by the camera service can pass an argument other than USE_CALLING_UID
to this method.

Bug: 27616192
Change-Id: Idb06112201b805a8b5c979b5f0761fec1c6994a3

Former-commit-id: 62deabcc4f15e9f15a031a05cb48665ae83d94e1