Grant Camera Permissions to DialerPackage
The Dialer apps now need camera permissions because they support video
calling. Having the app  ask for permission at runtime is not reasonable
because this is difficult to do in situations like:
  - incoming video calls
  - outging emergency video calls

It also would lead to a poor user experience in the scenario mentioned
above.

We already do this for a couple other permissions including microphone
permissions, and since the idea here is the same we are requesting camera
permissions for the dialer app as well to support video calling.

bug=30036076

Change-Id: Ie2c1bdbcb97236e8497339512dcdb8d921641cf6

Former-commit-id: a2b109a0a8f9866161d7031c06990dc5e471a709