Add the ability to postpone transitions.
Bug 30502032
Bug 20068723

Fragments in a transaction can now indicate when they are ready
and the transaction should complete. Postponed transactions will
have new views in the fragments created and added to the hierarchy
INVISIBLE. When all postponed fragments are ready, the transitions
and animations will start. INVISIBLE views will be made VISIBLE
and removed views will be taken out of the View hierarchy

Test: Id72daf16a62658227ee21a1aaf18a9b7657b6ba1

Change-Id: Ie42034629fee73299f8f82f44506aabf89970c3f

Former-commit-id: 86bfc666d33592f7db5584036aab8c44164d8dc2