Test existence of wallpaper files before backing them up
Otherwise we attempt to link a nonexistent file, which throws, and then
we abort the whole backup outright.  This way, we back up exactly what is
present regardless of the presence of each file individually.

We also now make sure to reset to the factory state and then apply the
restored content from there.

Bug 30102506

Change-Id: Ia6a39114f6b84e8cc01342df3da99833d7ca8e9b

Former-commit-id: 5f829d287f5c8cd72fee1522b8308fa459adc122