Suppress exceptions for corner-case a11y events.
Apps are told to query AccessibilityManager#isEnabled before
calling sendAccessibilityEvent. If accessibility is disabled
between the two calls, an app can crash. We can guarantee that
this won't happen on a process's main thread, but guaranteeing
it for all threads is messier. Rather than add the complexity
of tracking the state for all threads, only log an error in the
corner case that an accessibility event is requested on a
thread that doesn't have the process's main looper.

Bug: 28985452
Change-Id: I8369deefd83b0a6b04936ddfce55c53147756f1f

Former-commit-id: 890202bde486ca47ceb5daf4955038352384558e