docs: Updating image download pages for DP5
Updating with current images, which I've also uploaded to the cloud
share. I briefly made the cloud share public, so I could test the
download links and verify the checksums on the web page against the
downloaded files; the downloads and checksums all worked, so I made
the cloud files private again (so the links in these pages don't
work at the moment).

Also updated the download tables in the internationalized versions
of these files, since it's just a copy/paste.

See first comment for doc stage location.

bug: 29738757

Change-Id: I384786dcdf0dd249be0d1d0e38c88fa2b2109e0b

Former-commit-id: 5ba9c49964d039de0fbb383ab26685f5f205dfde