Also report resize when frame changes without inset change.
Currently we report resize via two main code paths:
    1. Insets change.
    2. Drag resizing/resized while not drag resizing.
Unfortunately the case of IME dismissal with SOFT_INPUT_ADJUST_RESIZE
will not trigger either. For #1, both the content and the parent frames
are adjusted together (similar to a docked resize), and so we won't
produce any insets beyond the system ones. For #2, we would only hit this
path if we went through the Task, but this all happens in PhoneWindowManager
layout. Prior to 3ccc5273 ("only resize during relayout"), the lack of
resize reporting wasn't a significant issue, as we would go ahead
and resize the dialog anyway. Assuming it wasn't in the middle of a
frame it would eventually catch up and render things correctly. Following
this change though we need to ensure we trigger the client calling relayout.
We accomplish this simply by also reporting frame changes.

Bug: 30191926
Change-Id: I95c7553e5e219e4a50c92f4d47621a32567a626f

Former-commit-id: 31aa98bea8b41e4e0e1f0c78be18dce8ef597f79