Attempt at fixing broken tests and regressing performance.
1. testDeleteDocument_Cancel seems to pass far more often than
testDeleteDocument, which always errors out due to not able to find the
menu item. This is to combine the tests so we don't lose coverage, but
also have less noise in APCT until we figure out what's going on.

2. We have regression currently on our jank tests. Per Guang's
suggestion, we are going to try freezing device orientation to see if it
helps.

Bug: 30190207
Change-Id: Ibdfdb75471b3220faaa9e8abef03df489ff185c3

Former-commit-id: 8c3e68d316b276f853fb484f9035e2c10e8b4a6e