DO NOT MERGE -- Correctly finish activity in non-focused stack
When activity is finished we first looked for next activity to
show in focused stack. If real next activity to show in place
of finishing one is in the same non-focused stack, we didn't
fully complete the dismissal process and activity was stuck in
FINISHING state.
This CL checks if we're trying to finish visible activity in
paused state and destroy it immediately if top running activity
is visible - same as we do for pinned activities.

Bug: 29458854
Change-Id: I0d5ceb2daa45c0628d89417c8456e132996bcea9
(cherry picked from commit d84977735a08c6fd9b46efe6573ea4c99563e147 [formerly 7318d63ba6dbb3042907d10d5369fcd5ac444d67])

Former-commit-id: 620040db3ff81a8318dfaee64dee4dae94c428fa