docs: Fixed bad link to cloud-backup training class
The name and URL of the old "Syncing to the Cloud" class had changed.
Fixed a single bad link; also changed the title on other pages where
the class was referred to by the old name.

I know that changing the _toc.cs files is redundant at this point
(since we don't use those files to generate TOCs any more) but I
figured I'd do it for consistency's sake.

See first comment for doc stage location.

bug: 30309375
Change-Id: I0cd8b1442ef3a8dd8a1917e897a3f0555c48e6ef

Former-commit-id: c02411f38d37c270460dc021770a1beab43a43c0