Enforce consistent sizes for arrays in SpannableStringInternal
The grow logic in SpannableStringInternal#setSpan assumes that the
size of mSpanData is consistent with that of mSpans, in particular
that if the latter doesn't need to grow, neither does the former.
The copySpans() method didn't enforce this, creating an mSpanData
array only big enough to hold the data.

This patch documents the invariant in a comment and enforces it.

Bug: 30359314
Change-Id: Ie25db70a76836e97af8476a7f5c10cb4b976c1cf
(cherry picked from commit b08a5f9a0232992c8ded51bb6dfe41ec3f608a2e [formerly 83549088c643586aa6013d947ad2c21464a3878e])

Former-commit-id: 9cde7244b633028d12fedcf3c66f16c0f25af9a7