docs: Removed reference to Android Support Library
In the Android Studio world, devs download the Android Support
*Repository*, not the now-deprecated *Library*. Updated the prose
and a screenshot to reflect this. Did other cleanup to docs to make
them more consistent with the new Android Studio/Gradle world. (For
example, the process for adding a library is now the same whether
or not the library has resources, so I consolidated the two "adding
a library" sections, and removed the paragraphs from features.html
that specified whether or not a library had resources.)

Also fixed bad formatting of code snippets (once again, we got bit by
the {@code &lt;} issue). Changed the {@code} tags to <code></code> to
fix it.

Other misc. fixes too, as they popped up in the review process.

See first comment for doc stage location.

bug: 30269620
Change-Id: I72922655e9887d1365e97b75982432c97e041f1f

Former-commit-id: 07fe016b84bcfc1ba2000824d6c32c34f5724df2