Only restore saved surfaces in the last visible rotation.
Sometimes we see a sequence like this when transitioning
between apps causes rotation changes.

1. New app becomes visible, causing the orientation to change
2. Old app gets marked as invisible, at this point its window
   are still visible from client, so surface is marked as eligible
   to be saved.
3. Old app calls relayout to invisible, we save surface.

Now if we restore the Surface it will be in the wrong orientation. It
never had a visible layout in the new one, it was just covered up
by the screenshot. To address this we track the rotation a window was
last visibily laid out in. If it doesn't match our current
rotation we can't use the saved surface.

Bug: 30171992
Change-Id: If814131810b6d01c5112afd145072a13a30b0401

Former-commit-id: 237028ae3be891dd1dbc392a51fd768474de5df7