Polish around locked user communication.
Based on user feedback, using a notification to communicate that a
user was locked was confusing.  Instead, this CL moves to showing a
message in the keyguard indicator slot at the bottom of the screen,
which is where we communicate other information such as charging
status.

Since a locked user is a very limiting experience, this new message
takes precedence over any other indicator message.

Bug: 28942360
Change-Id: I708d2b4b41079d09d46c17365c750488204e3090

Former-commit-id: b6edaa952eb04137c126f61122571f41087ad980