Try to resolve IDs also as enums
This is intended to cover the edge case in ConstraintLayout (and
possibly in other places) where an attribute is defined as
reference|enum.
If we can not resolve the value as a reference, try to resolve it as an
enum and return the value.

Change-Id: I2817aa5d78500247a2e9aec5411586a1db13791d
(cherry picked from commit 0ae4958dcb8e819021886874dc6e823da233565d [formerly b24b563654bf7c007f0912bf32fbab948fcb6daa])

Former-commit-id: f627cbb15b5ffc8e1195dfd1dde751a9202e11e1