Fix flickering issues with FLAG_SHOW_WHEN_LOCKED and fingerprint wake
The whole sequence is different as Keyguard is occluded. Thus, we are not
doing the normal keyguard exit sequence, but jump directly to dismissing
the Keyguard.

Normally the sequence is:
- keyguardDone
- notifyScreenTurningOn
- startKeyguardExitAnim

In the occluded case, it is:
- keyguardDone
- startKeyguardExitAnim
- notifyScreenTurningOn

Now, the following issues cause the flickering in the occluded case, which
are normally not causing any breakage:
- There was an issued with the draw callback not being reset and reused
at the wrong time.
- mWakeAndUnlocking was not cleared soon enough. This caused a timeout if
notifyScreenTurningOn happens after starKeyguardExitAnim
- We always need to wait for the Keyguard host window, as isForceHiding
isn't set in this case

Change-Id: Id8d0f9fec79ec63b36659513f97c724147c7521f
Fixes: 30483392

Former-commit-id: e93e6f9deb8d3c044d673ffb771b73b2346486a3