docs: Updating TIF docs for new TIF Companion library.
The TIF Companion library makes creating a TIF app much easier, and
will be the recommended starting point for devs going forward. This
update adds docs on using this library, and modifies the existing
docs for creating a TIF app from scratch (for devs that can't use
the companion library).

Bug: 30678049
Change-Id: Iaa5a6d0a4c9f0f2faf1e30549aba754ced7cc27a

Former-commit-id: e0810402fb3f687ccc8005608554ff7644ec951c