Allow to keep screen on only if window can be seen
If user has static wallpaper on lock screen and live wallpaper
on home screen, when screen is locked, wallpaper will not be an
obscuring window. In this case, if there is a non-obscured app
window behind the lock screen which has FLAG_KEEP_SCREEN_ON set,
it will not allow to turn screen off automatically, although it
is not really visible behind lock screen.

This CL restricts holding screen windows to only ones that can be
seen. Same applies for screen and button brightness and user
activity timeout settings.

Bug: 30359386
Change-Id: I46de831211c943d8077282e3274b2df180739239

Former-commit-id: b20addbdfb6e09f1721798db531aeec73d45a374