Allow seamless rotation before mTopIsFullscreen is set.
The difference between mTopFullscreenOpaqueWindowState != null
and mTopIsFullscreen relates to whether the status bar is animating.
However the flag won't be cleared until the next layout pass following
the animation. As long as the window isn't animating we are fine to
rotate seamlessly. This check was copied from selectRotationAnimationLw.
A little code archaeology implies it was perhaps historical and
introduced before the flag had this meaning.

Bug: 30171992
Change-Id: I326bf7766f8ebe307b833d1ca0c0cdfe80b1eb6c

Former-commit-id: 1ccd42510fe40f5ece98c54cc6fcd2b2724440ac