Request a wallpaper update when restoring surface of a wallpaper window
Normally wallpaper redo will be requested when the window draw finish.
If we're restoring, it's basically drawn instantly, and we need to update
so that the transition selection recognizes the correct wallpaper set up.

(Usually this scenario won't happen when going Home because home stack
don't save surface. However cts ActivityManagerTransitionSelectionTests
are simulating home with a task in fullscreen stack, and we might have
saved surfaces there.)

Change-Id: Ic5ba59ce54b279cc463888aefa94f7ddbf8a2bd3

Former-commit-id: 6e9872bd068188012d012e11d4445ac106360344