Add resource based mechanism to grant default permissions
This CL adds a mechanism for an OEM to grant default permissions
using the build system. This allows permissions for which they
got default grant exceptions to be added without modifying the
platform code. The format is a simple XML listing the packages
and the permissions to be granted for a package. The XML file
is placed in etc/default-permissions.

bug:29546655
bug:30929033

Change-Id: Ib6d671dac9201812bd2c47010c2c4b1330ef6b6e

Former-commit-id: d44d46aa507de39b7160b5e1358db9faad3da7e5