Hold no locks when calling RemotePrintSpooler
The calls might be blocking and need the main thread of to be unblocked
to finish. Hence do not call while holding any monitors that might need
to be acquired the main thread.

The calls that have been moved out of the lock:
- Icon related calls: These are just caches. Hence we flush the cache we
  just request the icon again
- Prune call: This just removes not installed print service from some
  data structure. The order of two of these calls does not matter as
  the end result of both removals will be the same, regardless of order.

Testing done: Reinstalled a print service multiple times. Before the
first reinstallation locked up the userState, now not anymore

Change-Id: I4f4cdaba65132dc2ef054877cbb097b499a723f6
Fixes: 31043684

Former-commit-id: 1d6e7cc536cb7f49b318f630ad8f2eb348786716