DO NOT MERGE ANYWHERE default_wallpaper: replace with black
The default_wallpaper is not displayed on Android Wear devices and it
can be removed to save space on the system partition. It is replaced
with black PNGs in the event that it is loaded and something in
frameworks expects these files to be present.

The black replacements are just a few kB compared to the few MB of the
original copies.

Bug: 28215770
Change-Id: Ie5327c5f32f16dad4bdc9c09ef8a1efa850f3ae8

Former-commit-id: d8ca8922e3d193f67acfadb0eaab3f85fa4e427a