Query geo description in worker thread to prevent ANR
Currently geo description of a caller is queried in the main thread, for
both InCallUI and framework CallerInfo. Since it is a potentially
time-consuming operation especially when the system is under heavy
loading, the main thread might be blocked for a long time. As a result,
the ANR happens.

The solution to to move the query to worker thread to free the main
thread and enhance the performance.

Besides, to ensure listeners are notified with complete CallInfo, we
need to delay callbacks until EVENT_END_OF_QUEUE.

Change-Id: Ibe444fc0f98d07e7ebbe266987e60ede334a3216

Former-commit-id: 3b90d48e81b34337a4a30f6a25970df9ec13b397