Revert "Dispatch a11y events in separate thread."
This reverts commit 2da277696d0ef7f8819ef7a3eddc423df3786e7c [formerly c34649411d053185b3572c4cd924e6f14295d8cd].

Dispatching accessibility events in their own thread is causing Chrome and gmail to crash. We've identified two issues: Chrome is allocating strings natively using references that aren't valid outside of their thread, and the text is being set to values that are changed in the UI thread. 

I'm going to resolve these issues on master by making deep copies of the strings, but that change will have its own performance implications.

Since we were bit almost immediately by an unexpected result of this change, and I need to erode its benefit by making deep copies, I think it's a bad bet to push it into MR1.

Bug: 31042124
Change-Id: I6f5c225a9197036db43fd0ac6008447b22617525

Former-commit-id: 5de2feab261831e2d6fbe28eab53b884860cd6ea