Redact work notifications separately from personal
This will strictly decrease the set of cases where work notifications
can be shown.

They will continue to honour the parent profile's settings, but will
also get redacted/hidden if the work profile has stricter settings
than the parent or when the parent is unlocked but the profile is still
locked.

Bug: 31001762
Change-Id: I2152631dbb8beb7a9899d6406b05f4447d757010

Former-commit-id: 92b83c6495abb3c091474b3a2f16dcbd949d6da7