Update JavaDoc for IMS#enableHardwareAcceleration
This makes it clear that on Android L+ devices IME developers basically
no longer need to call InputMethodService#enableHardwareAcceleration(),
because it is deprecated in Android L[1][2].  On L+ devices what IME
developers need to do is to make sure that their IMEs are compatible
with both hardware and software rendering so that the system can choose
the best rendering method.

This is basically a code clean-up that should not change any actual
behavior.

 [1]: I906d48cb07af50fa02dae4f4ecdb5e65211fc6ef
      677436825354e0ec43b844a941f071bd33f0e5c1 [formerly 5e1565ead6dbb7d5c414522591f61b16a23de1c3]
 [2]: I126ea823eb2eeabb29d30999dec9c9411b33b315
      bb8aff64fa93b761a83c3e74c6684e97808c715c [formerly e07b595519e68a8ee524b1886dea01475047abec]

Bug: 31202732
Change-Id: I416669b172737e3eb3d77165dcea50289d207514

Former-commit-id: ac0211a47801dd8476f48dec7535b2136a324a19