Add editable_voicemail_number key to CarrierConfig
Add KEY_EDITABLE_VOICEMAIL_NUMBER_BOOL and default value to
CarrierConfigManager, and remove the references to the old config in
core/res/res/values/[config.xml/symbols.xml]

This value exists to fix http://b/14093735. If a SIM card has no
voicemail number available, since the default number is empty, we allow
users to edit the voicemail number when this flag is true.

Bug: 29873049
Change-Id: I68a494e2176306e4e9181a8ccb0aa818f3e08154

Former-commit-id: 4a54638a6829c535f18003618abe729c38d7dbc1