Add LatencyTracker to track important SystemUI latencies for testing
Add a new class that can record latencies for various actions. Add a
new system property debug.systemui.latency_tracking to enable/disable
latency as well as a broadcast to reload the property if it changes.

This will be used to write lab latency tests to track important
transitions/latencies.

Change-Id: I11f96f1c410d83f2d3dbbdf903b29e3035744fb8

Former-commit-id: e05256e2a1edf6f57f9fb48ef4f98dc4261ab8a7