[RingtoneManager] API to add custom ringtones
So far, the only way for users to add custom ringtones was to either use
a 3rd party app or copy files to a specific directory in the storage,
and there was no API to make it easier to add custom ringtones.

This CL creates a hidden API in RingtoneManager that allows adding any
audio file as a custom ringtone, by copying that file into the ringtone
storage and asking MediaScanner to scan that file.

This API is not intended to be used by existing apps, but an associated CL
uses it in RingtonePickerActivity

Test: make && vendor/google/tools/flashall, plus manual testing with associated CL
Bug: 30658854
Change-Id: I2b291847c5430e1fc2c3fc6a351b6111a49fcbf3

Former-commit-id: 7c8ccb3ce4943a0e69db536058b97e8c0ffebb71