Reset mCurrentUserId when fingerprintd died
A template fingerprint data is not loaded in fingerprintd when
fingerprintd is died and restarted.
FigerprintService tries to set userId after a restart of fingerprintd,
but it's skipped because the service tries to set same userId already
set to mCurrentUserId.

To load correct template fingerprint data, reset mCurrentUserId
when fingerprintd died.

Bug: 31734514
Change-Id: I0d3452fc6c1f82bb2a2af10f61d3f3a47df9dec6

Former-commit-id: 895da77aa6baa9231faa2a1159b673b6c67db3b9